moved mapping classes and corresponding tests to specific tool folders - bt2gh and gh2bt
