--- src/bridge/pipelines/mapping/__init__.py
@@ -1,9 +1,0 @@
-"""
-Classes for mapping metadata records from one schema to another.
-"""
-
-from .bt2gh_map import MapBioTools2GitHub
-
-__all__ = [
-    "MapBioTools2GitHub",
-]
